requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
pydantic>=2.6.0
pydantic-settings>=2.2.1
python-dotenv>=1.0.1
pandas>=2.2.2
pyarrow>=16.1.0
requests>=2.32.3
